@@ -1,20 +1,19 @@
 ﻿using nGantt.PeriodSplitter;
 using System.Windows.Input;
 
-namespace nGantt.GanttChart
+namespace nGantt.GanttChart;
+
+public delegate void ContextMenuItemClick(GanttTask ganttTask);
+
+public class ContextMenuItem
 {
-    public delegate void ContextMenuItemClick(GanttTask ganttTask);
-
-    public class ContextMenuItem
+    public ContextMenuItem(ContextMenuItemClick contextMenuItemClick, string name)
     {
-        public ContextMenuItem(ContextMenuItemClick contextMenuItemClick, string name)
-        {
-            //   ContextMenuItemClickCommand = new DelegateCommand<GanttTask>(x => contextMenuItemClick(x));
-            this.Name = name;
-        }
-
-        public string Name { get; set; }
-
-        public ICommand ContextMenuItemClickCommand { get; private set; }
+        // ContextMenuItemClickCommand = new DelegateCommand<GanttTask>(x => contextMenuItemClick(x));
+        this.Name = name;
     }
+
+    public string Name { get; set; }
+
+    public ICommand ContextMenuItemClickCommand { get; private set; }
 }
