@@ -3,10 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace nGantt.GanttChart
-{
-    public class GanttRowHeader
-    {
-        public string Name { get; set; }
-    }
-}
+namespace nGantt.GanttChart;
+
+public record GanttRowHeader(string Name);
