@@ -4,17 +4,16 @@ using System.Linq;
 using System.Text;
 using System.Windows.Media;
 
-namespace nGantt.GanttChart
+namespace nGantt.GanttChart;
+
+public class TimeLineItem
 {
-    public class TimeLineItem 
+    public TimeLineItem()
     {
-        public TimeLineItem()
-        {
-            BackgroundColor = new SolidColorBrush(Colors.Transparent);
-        }
-        public DateTime Start { get; set; }
-        public DateTime End { get; set; }
-        public string Name { get; set; }
-        public Brush BackgroundColor { get; set; }
+        BackgroundColor = new SolidColorBrush(Colors.Transparent);
     }
+    public DateTime Start { get; set; }
+    public DateTime End { get; set; }
+    public string Name { get; set; }
+    public Brush BackgroundColor { get; set; }
 }
