@@ -46,7 +46,7 @@ public partial class App : Application
 
 		// Prevent default unhandled exception processing
 
-		MessageBox.Show($"Unhandled Exception - please send screen shot to Maxwell if this wasn't supposed to happen:\n {e.ExceptionObject}",
+		MessageBox.Show($"Unhandled Exception\n {e.ExceptionObject}",
 			"Unhandled Exception", MessageBoxButton.OK, MessageBoxImage.Error, MessageBoxResult.None);
 	}
 }
